Add metadata loading from yaml header
Pandoc and others nowadays support yaml headers in markdown:
https://pandoc.org/MANUAL.html#metadata-variables

We also can put our own stuff there with them just ignoring it - so
switching from a custom header to that makes sense.

ModTemplate still supports the old {: :} syntax if it is the first
thing in the file, and can load yaml headers as well (overriding
content from the legacy header, if specified) if it directly follows
the legacy header. Otherwise the yaml header also is expected at the
beginning of the file.