blog/fs: allow autogenerating cdb if cdbmaxage is set
cdbmaxage takes a value in seconds. If the mtime of the cdb file +
maxage is less than the current time the directory tree will be
rescanned. The cdb file will only be generated if any of the index
files is newer than the cdb file, though. To avoid needless rescanning
the mtime of the cdb file is updated in either case.